Move PolymorphicRdata.
